Initial support for kernel ABI emulator
This initial code is not fully functional yet. Once it is finished, it
will allow to test latest BlueZ code without an up-to-date kernel. If
enabled, any AF_BLUETOOTH socket communication is intercepted by a
Valgrind wrapper library (bt-kernel.so), and sent to the testcase
process where it can be processed.

The mechanism is very similar to how HCI packets are encoded/decoded by
the /dev/vhci dispatcher, except that everything is done in userspace.
It is possible to even run BlueZ as non-root (by using a private D-Bus
bus and wrapping functions that require root privilege).